+++ Pix/view/frontend/web/js/view/payment/method-renderer/openpix-boleto-method.js
@@ -1,0 +1,29 @@
+define([
+  'Magento_Checkout/js/view/payment/default',
+  'jquery',
+  'Magento_Checkout/js/model/payment/additional-validators',
+  'Magento_Ui/js/model/messageList',
+], function (Component, $, validators, messageList) {
+  'use strict';
+  return Component.extend({
+    defaults: {
+      template: 'OpenPix_Pix/payment/openpix_boleto',
+    },
+
+    initObservable: function () {
+      this._super().observe(['cpfCnpj']);
+
+      return this;
+    },
+
+    getData: function () {
+      return {
+        method: this.item.method,
+      };
+    },
+    getCode: function () {
+      return 'openpix_boleto';
+    },
+    // add required logic here
+  });
+});