+++ Pix/view/frontend/web/js/view/payment/openpix_boleto.js
@@ -4,9 +4,9 @@ define([
 ], function (Component, rendererList) {
   'use strict';
   rendererList.push({
-    type: 'openpix_pix_parcelado',
+    type: 'openpix_boleto',
     component:
-      'OpenPix_Pix/js/view/payment/method-renderer/openpix-pix-parcelado-method',
+      'OpenPix_Pix/js/view/payment/method-renderer/openpix-boleto-method',
   });
   /** Add view logic here if needed */
   return Component.extend({});